+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removed some unneded and wrong placed logging messages
 - The AzuraCast Settings embed is now shown even when the instance is offline
   - The AzuraCast Stations embed however continues to be not shown
+- When the AzuraCast instance has a self-signed SSL certificate Azzy will now warn about it
 
 ## 2.2.3 - 2025-01-15
 ### Dependencies